[refactor] ProjectModal, ProjectCard 파일 이름 수정 - 소문자로 시작한 파일 이름을 대문자로 시작하도록 수정
